fix Communicator, XD: XSS patch. remove the parameterized swf= attribute in relay.html
Change the relay.html file and Communicator to no longer pass along the relay.swf location
as a url parameter. Previously, this parameter was passed through, creating a potential
security vulnerability when a malicious swf was set to the swf= parameter. This change
requires relay.swf to be in the same location as relay.html. To prevent spoofing, the
relay.swf file is a regex change of the relay.html's location.href property.

BREAKING CHANGE: the swf file previously could be located wherever the library user wished
and must now be located alongside the relay.html file.

  Before:
  |-myDir
    |-relay.html
  |-otherDir
    |-relay.swf

  After:
  |-myDir
    |-relay.html
    |-relay.swf

If an implementor must absolutely specify an alternate SWF location due to their environment,
they should modify their relay.html file and specify the ALTERNATE_SWF_LOCATION which has
been provided as a backwards compatible solution.