fix(RulesEngine): Adds support for calling define() with no base URL
In some scenarios, it's possible to call define() without the base URL
being defined. In these scenarios, we were throwing an exception
that we couldn't proceed further. This wasn't entirely true. This change
uses the location.pathname property as a fallback, and reduces the error
to a warning.

This allows named AMD modules to function by including them via a script
tag, even though a module root hasn't been configured.