fix(plugins) plugins now work with instances of Inject
the extras for inject surrounding the non AMD plugins were never modified to
work with the new instance based designs. This change allows a global
INJECT_PLUGINS object to exist, which can then be plugged in to.

BREAKING CHANGE
The prior inject-plugins.js file is no longer a part of recent, and the
plugins need to be included indivudally. Additionally, a plugin needs to be
enabled by passing the instance of Inject to the plugin method:

INJECT_PLUGINS.css(Inject); // adds the CSS plugin to the current instance