fix (Executor): Improves stacktrace-js integration
This fixes the stacktrace.js integration in two important ways. First, the
library is now beign included externally (much like Fiber). Second, the stack
normalization guards against invalid or unparsable traces (noticed in IE8
compatibility mode).

As part of the migration to the external stacktrace-js library, the tests were
updated to reference their node_module version. This will ensure for both
stacktrace-js and Fiber we are referencing the proper version that is being
built with during our unit tests.

Fixes #301