fix (Analyzer): Resolves overzealous matching of define statements
This resolves a problem with define() calls that did not declare a dependency
array, but had an array immediately defined at the top of their factory
function. In these cases, the array was being identified as dependencies. To
resolve this, the Inject reject for define() now scans for [^)] instead of
the prior [\w\W] pattern.